Move Components from `pages` to `components`.
We have accidentally started developing by adding React components to the `pages` directory, which should only hold larger, page-length components. This change moves the features to their proper location.
