+++ client/src/components/LoadingSuggestions.js
