+++ client/src/components/OriginalSentence.js
