+++ client/src/components/SuggestionSentence.js
